+++ plugins/helpers/common.js
@@ -1,0 +1,7 @@
+'use strict';
+
+const redis = require('../db/redis');
+
+module.exports = {
+	redis
+};
